+++ core/components/__tests__/horizontal-rule.test.js
@@ -80,4 +80,79 @@ describe('Markdown shorthand works', () => {
       </doc>,
     );
   });
+
+  it('type --- above a paragraph', () => {
+    const { view } = testEditor(
+      <doc>
+        <para>[]</para>
+        <para>test</para>
+      </doc>,
+    );
+
+    typeText(view, '---');
+    expect(view.state).toEqualDocAndSelection(
+      <doc>
+        <hr />
+        <para>[]test</para>
+      </doc>,
+    );
+  });
+
+  it('type --- above a hr', () => {
+    const { view } = testEditor(
+      <doc>
+        <para>[]</para>
+        <hr />
+      </doc>,
+    );
+
+    typeText(view, '---');
+    expect(view.state).toEqualDocAndSelection(
+      <doc>
+        <hr />
+        <para>[]</para>
+        <hr />
+      </doc>,
+    );
+  });
+
+  it('type --- inisde a nested blockquote', () => {
+    const { view } = testEditor(
+      <doc>
+        <blockquote>
+          <blockquote>
+            <para>[]</para>
+          </blockquote>
+        </blockquote>
+      </doc>,
+    );
+
+    typeText(view, '---');
+    expect(view.state).toEqualDocAndSelection(
+      <doc>
+        <blockquote>
+          <blockquote>
+            <hr />
+            <para>[]</para>
+          </blockquote>
+        </blockquote>
+      </doc>,
+    );
+  });
+
+  it('type --- inisde a paragraph with more text', () => {
+    const { view } = testEditor(
+      <doc>
+        <para>[]abc</para>
+      </doc>,
+    );
+
+    typeText(view, '---');
+    expect(view.state).toEqualDocAndSelection(
+      <doc>
+        <hr />
+        <para>[]abc</para>
+      </doc>,
+    );
+  });
 });